Added comments to the VarableDeinfer class and renamed the class to VariableMgmt. Started the boilerplate code for the Sentence Builder class
